+++ src/main/scala/fr/janalyse/primes/Shutdownable.scala
@@ -1,0 +1,5 @@
+package fr.janalyse.primes
+
+trait Shutdownable {
+  def shutdown():Unit
+}